support ColorOS!!! fix bugs

+++ app/src/main/java/com/zhizi42/diymiuicard/Utils.java
@@ -2,7 +2,7 @@ package com.zhizi42.diymiuicard;
 
 import android.app.Activity;
 import android.content.Context;
-import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
 import android.util.Log;
 import android.widget.Toast;
 
@@ -12,7 +12,6 @@ import com.bumptech.glide.Glide;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
-import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
@@ -20,16 +19,17 @@ import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.zhizi42.diymiuicard.Hook;
+
 public class Utils {
 
-    public static SharedPreferences sharedPreferences;
+    public static boolean debug;
 
-    public static void initSetting(SharedPreferences sharedPreferences0) {
-        sharedPreferences = sharedPreferences0;
+    public static void setDebug(boolean debug0) {
+        debug = debug0;
     }
 
     public static boolean executeShell(List<String> cmd) {
-        boolean debug = sharedPreferences.getBoolean("debug", false);
         try {
             Process process = Runtime.getRuntime().exec("su --mount-master");//加上--mount-master才能读写/data/data下其它应用
             BufferedWriter writer = new BufferedWriter(
@@ -81,29 +81,58 @@ public class Utils {
                 .setPositiveButton(R.string.confirm, (dialog, which) -> {
                     Glide.get(activity).clearMemory();
                     new Thread(() -> {
+                        int OSType = checkOSType(activity);
                         Glide.get(activity).clearDiskCache();
                         List<String> commandList = new ArrayList<>();
                         commandList.add("rm -rf /data/data/com.zhizi42.diymiuicard/cache/*");
-                        commandList.add("am force-stop com.miui.tsmclient");
-                        commandList.add("rm -rf /data/data/com.miui.tsmclient/cache/image_manager_disk_cache/*");
+                        if (OSType == 0) {
+                            commandList.add("am force-stop com.miui.tsmclient");
+                            commandList.add("rm -rf /data/data/com.miui.tsmclient/cache/image_manager_disk_cache/*");
+                        }
                         boolean succ = Utils.executeShell(commandList);
                         activity.runOnUiThread(() -> {
                             if (succ) {
-                                Toast.makeText(
-                                                activity,
-                                                "清除缓存成功",
+                                Toast.makeText(activity,
+                                                R.string.clear_card_cache_succ,
                                                 Toast.LENGTH_SHORT)
                                         .show();
                             } else {
-                                Toast.makeText(
-                                                activity,
-                                                "清除缓存失败，请手动清除",
+                                Toast.makeText(activity,
+                                                R.string.clear_card_cache_fail,
                                                 Toast.LENGTH_LONG)
                                         .show();
                             }
                         });
                     }).start();
                 })
+                .show();
+    }
+
+    public static void debugLog(Hook hook, String s) {
+        if (debug) {
+            hook.log(s);
+        }
+    }
+
+    public static int checkOSType(Context context) {
+        try {
+            context.getPackageManager().getPackageInfo("com.miui.tsmclient", PackageManager.GET_META_DATA);
+            return 0;
+        } catch (PackageManager.NameNotFoundException e) {
+            try {
+                context.getPackageManager().getPackageInfo("com.finshell.wallet", PackageManager.GET_META_DATA);
+                return 1;
+            } catch (PackageManager.NameNotFoundException ex) {
+                return -1;
+            }
+        }
+    }
+
+    public static void showNoCardApp(Context context) {
+        new AlertDialog.Builder(context)
+                .setTitle(R.string.no_card_app_installed_title)
+                .setMessage(R.string.no_card_app_installed_content)
+                .setPositiveButton(R.string.confirm, null)
                 .show();
     }
 }
